+++ www/whatsnew.md
@@ -11,6 +11,7 @@ permalink: /whats-new/
  - Necronomicon is now an Occultist card like it was supposed to be. (1523)
  - Forge Suit now properly only gives Health and not Attack. (1525)
  - Sand Filter now actually Decays, and is properly now only active on your own turn. (1525)
+ - Cosmic Apparitions now correctly only discovers minions and also adds a copy to your hand. (1534)
  
 In addition, extensive edits by @Logovaz summarized here: